+++ server/web/session/redis/sess_redis_test.go
@@ -15,21 +15,22 @@ import (
 )
 
 func TestRedis(t *testing.T) {
-	sessionConfig := &session.ManagerConfig{
-		CookieName:      "gosessionid",
-		EnableSetCookie: true,
-		Gclifetime:      3600,
-		Maxlifetime:     3600,
-		Secure:          false,
-		CookieLifeTime:  3600,
-	}
-
 	redisAddr := os.Getenv("REDIS_ADDR")
 	if redisAddr == "" {
 		redisAddr = "127.0.0.1:6379"
 	}
+	redisConfig := fmt.Sprintf("%s,100,,0,30", redisAddr)
 
-	sessionConfig.ProviderConfig = fmt.Sprintf("%s,100,,0,30", redisAddr)
+	sessionConfig := session.NewManagerConfig(
+		session.CfgCookieName(`gosessionid`),
+		session.CfgSetCookie(true),
+		session.CfgGcLifeTime(3600),
+		session.CfgMaxLifeTime(3600),
+		session.CfgSecure(false),
+		session.CfgCookieLifeTime(3600),
+		session.CfgProviderConfig(redisConfig),
+	)
+
 	globalSession, err := session.NewManager("redis", sessionConfig)
 	if err != nil {
 		t.Fatal("could not create manager:", err)